requirements.txt
requests==2.31.0
click==8.1.3
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
python-dotenv==1.0.1
bs4==0.0.2